Add achievement badges, app notification counters, activity type labels and weight/BP/glucose trends; split nxk parser and Dashboard into modules
- Decode Notify's achievement badges (ach-* keys in appSetting) into a dedicated Overview section, never touching unrelated settings or third-party tokens stored in the same table
- Surface statsApp's per-application counters alongside the existing daily notification breakdown
- Replace opaque record/workout type codes with honest, non-invented labels (workout title when present, otherwise a clearly flagged raw code)
- Add weight, blood pressure and glucose to the multi-day comparison metrics, aggregated per day in buildDays
- Fix a bug where day-level spo2Average always fell back to 0 because the aggregate's spo2Count was never initialized
- Split src/lib/nxk.js (657 lines) into src/lib/nxk/{sqlHelpers,schema,geo,records,days,index}.js
- Split src/components/Dashboard.jsx (562 lines) into src/components/dashboard/{shared,Overview,ArchiveOverview,SleepView,HeartView,VitalsView,ActivityView,index}.jsx
- Add unit tests for the new pure modules (achievements, appStats, activityTypes, days)
