+++ src/styles.css
@@ -442,6 +442,14 @@ p { text-wrap: pretty; }
 .fact-grid dd { margin: 0; overflow-wrap: anywhere; font-size: 0.87rem; font-weight: 760; }
 .fact-grid small { display: block; margin-top: 3px; color: var(--muted); font-size: 0.68rem; font-weight: 500; }
 
+.achievement-grid { display: grid; grid-template-columns: 1fr; gap: 10px; }
+.achievement-card { display: grid; grid-template-columns: auto minmax(0, 1fr); gap: 12px; padding: 14px; border-radius: 12px; background: var(--surface); }
+.achievement-card__icon { display: grid; flex: 0 0 auto; place-items: center; width: 36px; height: 36px; border-radius: 10px; color: var(--primary); background: var(--primary-soft); }
+.achievement-card strong { display: block; font-size: 0.9rem; }
+.achievement-card__category { display: inline-block; margin-top: 3px; color: var(--accent); font-size: 0.68rem; font-weight: 760; text-transform: uppercase; letter-spacing: 0.04em; }
+.achievement-card p { margin: 6px 0 0; color: var(--muted); font-size: 0.8rem; line-height: 1.45; }
+.achievement-card small { display: block; margin-top: 6px; color: var(--muted); font-size: 0.7rem; }
+
 .detail-block { min-width: 0; margin-top: 20px; }
 .detail-block + .detail-block { padding-top: 18px; border-top: 1px solid var(--line); }
 .detail-block__header { display: flex; align-items: flex-start; justify-content: space-between; gap: 12px; margin-bottom: 11px; }
@@ -712,6 +720,7 @@ p { text-wrap: pretty; }
   .session-row span { margin-bottom: 5px; }
   .gps-summary { grid-template-columns: repeat(4, 1fr); }
   .fact-grid { grid-template-columns: repeat(4, minmax(0, 1fr)); }
+  .achievement-grid { grid-template-columns: repeat(2, minmax(0, 1fr)); }
   .quality-summary { grid-template-columns: repeat(4, 1fr); }
   .comparison-controls { flex-direction: row; align-items: end; justify-content: space-between; }
   .comparison-controls label { min-width: 210px; }
